Map Updated
Right now, the white portion routes to Map. I feel this should be ProShows.
The Map now has clickable buttons at icsr, gc, sac, clt, oat, lib.
Because of current state of db, clicks to sac, clt and oat give stalls, but should work if database is completed.
If times of events are unknown at time of app release, please modify onCreateContextMenu to load three randomized events or three highest profile events (ask cores) at the venue. This has not been implemented. Right now, it assumes time info is available and shows the current and upcoming events, with Day 1 hardcoded (MapActivity.java:264) instead of current day.
The place to redirect to location-based events activity is on MapActivity.java:439.
The imagebuttons look rather ugly. Please modify in the resources if needed.
I tried tying the map with resources but because of the handling further down this would mean high load time for map, hence sticking to sdcard. (File vs InputStream; MapActivity.java:183)
Aside, can the QR code for the app be merged with the logo or something when the app is released? That way, during saarang ppl can download the app on the go...